Fix GetCustomLeaderboardById endpoint

+++ DevilDaggersWebsite.BlazorWasm/Shared/Dto/Public/CustomEntries/GetCustomEntry.cs
@@ -6,7 +6,7 @@ namespace DevilDaggersWebsite.BlazorWasm.Shared.Dto.Public.CustomEntries
 	{
 		public int PlayerId { get; init; }
 		public string PlayerName { get; init; } = null!;
-		public int Time { get; init; }
+		public double Time { get; init; }
 		public int GemsCollected { get; init; }
 		public int EnemiesKilled { get; init; }
 		public int DaggersFired { get; init; }
@@ -18,9 +18,9 @@ namespace DevilDaggersWebsite.BlazorWasm.Shared.Dto.Public.CustomEntries
 		public int GemsEaten { get; init; }
 		public int GemsTotal { get; init; }
 		public byte DeathType { get; init; }
-		public int LevelUpTime2 { get; init; }
-		public int LevelUpTime3 { get; init; }
-		public int LevelUpTime4 { get; init; }
+		public double LevelUpTime2 { get; init; }
+		public double LevelUpTime3 { get; init; }
+		public double LevelUpTime4 { get; init; }
 		public DateTime SubmitDate { get; init; }
 		public string? ClientVersion { get; init; }
 	}
